Wayland: Add new display server skeleton
Add the basic elements for a new display server using the 'dummy'
renderer (which does nothing). This patch only prepares for the Wayland
support and is not functional.

To test it so far:
* Start a wayland based compositor (e.g. Weston)
* Start Godot pasing '--display-driver wayland'

If it works, you won't see any windows. If it fails, you should see
error messages in the stderr.